% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{allGeneOut}
\alias{allGeneOut}
\title{Return a table of C,K GRanges with all C,K (with duplicates if they exist)}
\usage{
allGeneOut(digest, gene)
}
\description{
Return a table of C,K GRanges with all C,K (with duplicates if they exist)
}
