% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cin.R
\name{.cinFG}
\alias{.cinFG}
\title{Computes the number of focal gains}
\usage{
.cinFG(DIGEST, PLOIDY)
}
\arguments{
\item{DIGEST}{digest/vault$tables$segment}
}
\value{
integer of focal gains
}
\description{
Computes the number of focal gains
}
