% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{geneOut}
\alias{geneOut}
\title{Return a genes x C,K GRanges with most exotic C,K chosen for each gene}
\usage{
geneOut(digest, gene)
}
\description{
Return a genes x C,K GRanges with most exotic C,K chosen for each gene
}
