% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{geneTileOut}
\alias{geneTileOut}
\title{Return deduplicated Gene, C and K from all tiles (only covered genes)}
\usage{
geneTileOut(digest, gene)
}
\description{
Return deduplicated Gene, C and K from all tiles (only covered genes)
}
