% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gg.R
\name{ggArrangedPlot}
\alias{ggArrangedPlot}
\title{Plots summary plot for model picking}
\usage{
ggArrangedPlot(dig, genes, gg.opts)
}
\arguments{
\item{dig}{a digest exported from cnvex}

\item{genes}{a granges object}

\item{gg.opts}{obtained from getGgOpts()}
}
\description{
Plots summary plot for model picking
}
